pyproject.toml
[project]
name = "poc-vector-sqlite"
version = "0.1.0"
description = ""
authors = [
    {name = "VladimirMonin",email = "vladimir.monin2016@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "peewee (>=3.18.3,<4.0.0)",
    "sqlite-vec (>=0.1.6,<0.2.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "numpy (>=2.3.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "markdown-it-py (>=3.0.0,<4.0.0)",
    "google-genai (>=1.0.0,<2.0.0)"
]

[project.optional-dependencies]
media = [
    "Pillow (>=10.0.0,<11.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
pythonpath = ["."]
